requirements.txt
algoliasearch-django>=2.0,<3.0
django>=4.0.0,<4.1.0
djangorestframework~=3.14.0
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
pymongo~=3.12.3

certifi~=2022.12.7